VIDCS-3721: improvements-documentation-video-example
Removed the default * (catch-all) route for unfound API endpoints. This behavior was not API-friendly and caused confusion.

Fixed project scripts: some were intended to run both the backend and frontend simultaneously, but Bash is not concurrent by default. We now use concurrently (via npm) to achieve the desired result.

Added a script for debugging the backend project.

Updated the README to make the tutorial more detailed and beginner-friendly.

Readme now includes a clearer explanation of where and how to retrieve the necessary Vonage credentials.

Added steps for testing the app with multiple devices on a local network.